+++ CHANGELOG.md
@@ -15,6 +15,10 @@
     - Yields all body content first followed by all footnotes
     - Improved parsing speed
 
+### Quotation detection
+
+- Embeddings now are cached locally in `.remarx_cache/embeddings/` directory to improve performance.
+
 ## [0.3.0] - 2025-10-27
 
 ### Sentence corpus creation